Use UTF-16 <-> UTF-8 codecvt on Windows
Windows uses a 2-Byte `wchar_t` which effectively is UCS-2,
i.e. only codepoints up to U+FFFF.
So decoding an UTF-8 file with codepoints higher than that will fail.

As most Windows APIs now accept UTF-16 encoded `wchar_t` it makes sense
to also use that for the `codecvt` generated by the std-backend.

This is actually disallowed to be used for `std::basic_filebuf` but it
works in practice, so use it anyway.

Closes #71