Fix configure check for external iconv and ICU (64 bit)
Partial revert of 72a8e9eb75bd578b60d560f6c1f798af9cdfbf49
The `exe` rule implicitly creates an `obj` rule for the given cpp file if it doesn't exist yet. This causes trouble when the cpp file is meant to be built with different options by multiple `exe` rules as the 2nd rule will reuse the `obj` rule from the first and hence ignores the different options such as the include path.
Here it affects the `has_external_iconv` target due to the conflict with the `has_iconv` target.
Hence add back an `obj` rule for the former. Fixes #63
The same problem exists with the `has_icu*` targets. Fix it similarly.