Mark virtual classes as VISIBLE
On at least libc++ UBSAN has trouble detecting that the classes are
actually the same when (only) marked with `BOOST_LOCALE_DECL` because on
Linux/Clang when compiling the source it will evaluate to nothing hiding
the VTABLE which leads to:
> `runtime error: downcast of address 0x... which does not point to an object of type ...`

As we only need the import/export for the `std::locale::id` put that
into a separate class template which is marked `BOOST_LOCALE_DECL` using
CRTP to create unique IDs.
Then the actual classes can be marked `BOOST_VISIBLE` to create unique
VTables.