Add `time_put_from_base` back
This was removed as it is basically forwarding the call and hence looks
very similar to what `std::time_put_by_name` does.
However the `ios_base&` in the `do_put` function is actually used to
e.g. translate weekday names or determine the date format to use
(e.g. whether the year has 2 or 4 digits
or month comes before or after the days: dd/mm vs mm/dd)
Very likely the `std::locale::name` is used for that so we need to pass
a `std::ios_base&` with a locale that was constructed by the requested name.