Raise exception when iconv wrongly returns E2BIG
E2BIG shall be returned if the output buffer is too small.
We pass a 64 byte buffer which should be enough to do *something*

However it is observed on MacOS Sonoma that iconv returns E2BIG without
doing any progress leading to an infinite loop as the same inputs will
be passed over and over.
See #206.
Instead raise an exception when this case is detected.