Iconv: Skip invalid output chars in iconv if requested
If `skip` is selected we can use the `//IGNORE` suffix to ignore any
"character [that] cannot be represented in the target character set" in
iconv instead of doing so in the convert code causing (at least) multiple
invocations of `iconv`.
If the suffix is not supported (it is by libc and libiconv) fall back to
previous code.