Set ACTIONS_ALLOW_USE_UNSECURE_NODE_VERSION
Node 20 isn't supported in containers and GitHub actions forces actions
to use Node 20 even when the action is specified to use Node 16.
See https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default

This causes e.g. `actions/checkout@v3` to fail since that change.
A workaround is to set `$ACTIONS_ALLOW_USE_UNSECURE_NODE_VERSION` to
prevent this change (for now).